chore: lib structure
